# swagger.yaml
openapi: 3.0.0
info:
  description: |
    U ovom dokumentu su navedeni API-ji koje nudi modul profesori.
  version: "1.0.0-oas3"
  title: Mikroservisi modula profesori
...
paths:
...
  '/fox/teme/{idPredmeta}':
    get:
      tags:
        - teme
      summary: Pronadji teme za predmet
      description: Vraća sve teme na navedenom predmetu.
      operationId: getTeme
      parameters:
        - name: idPredmeta
          in: path
          description: id predmeta
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lista tema
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tema'
        '400':
          description: Invalid ID supplied
        '404':
          description: Not found
  /fox/teme/novaTema:
    post:
      tags:
        - teme
      summary: Upisuje se nova tema sa opisom
      operationId: addTema
      responses:
        '200':
          description: OK
        '400':
          description: Bad request
        '404':
          description: Not found
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NovaTema'
        description: Vrijednost koja se unosi za ocjenu studentu
        required: true
  '/fox/teme/izmjeniTemu/{idTeme}':
    put:
      tags:
        - teme
      summary: Izmjeni temu
      description: Mijenja temu sa id-em idTeme
      operationId: updateTema
      parameters:
        - name: idTeme
          in: path
          description: id teme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
        '400':
          description: Invalid ID supplied
        '404':
          description: Not found
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeme'
        description: Vrijednost koja se unosi za ocjenu studentu
        required: true
  '/fox/teme/izbrisiTemu/{idTeme}':
    delete:
      tags:
        - teme
      summary: Izbrisi temu
      description: Brise temu sa id-em idTeme
      operationId: deleteTema
      parameters:
        - name: idTeme
          in: path
          description: id teme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
        '400':
          description: Invalid ID supplied
        '404':
          description: Not found
...
servers:
  - url: 'https://virtserver.swaggerhub.com/dsabovic1/ProfesoriMikroservisi/1.0.0'
  - url: 'http://virtserver.swaggerhub.com/dsabovic1/ProfesoriMikroservisi/1.0.0'
components:
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: 'http://petstore.swagger.io/oauth/dialog'
          scopes:
            'write:pets': modify pets in your account
            'read:pets': read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header
  schemas:
...
    Tema:
      type: object
      properties:
        id:
          type: integer
          format: int64
        naziv:
          type: string
        opis:
          type: string
...
    NovaTema:
      type: object
      properties:
        naziv:
          type: string
        opis:
          type: string
        idProfesora:
          type: integer
          format: int64
        idPredmeta:
          type: integer
          format: int64
    UpdateTeme:
      type: object
      properties:
        naziv:
          type: string
        opis:
          type: string